# Documents/SmartCliff/EMS_BACKEND/docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
...
        "/events/create_new_event": {
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Create New Event",
                "operationId": "create_new_event_events_create_new_event_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EventBase"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EventResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/events/update_banner": {
            "patch": {
                "tags": [
                    "Events"
                ],
                "summary": "Update Banner",
                "operationId": "update_banner_events_update_banner_patch",
                "parameters": [
                    {
                        "name": "eventId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "description": "Event Id",
                            "title": "Eventid"
                        },
                        "description": "Event Id"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_update_banner_events_update_banner_patch"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/events/get_reschedule_dates": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get Possible Reschedule Dates",
                "operationId": "get_possible_reschedule_dates_events_get_reschedule_dates_get",
                "parameters": [
                    {
                        "name": "eventId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "description": "Event Id",
                            "title": "Eventid"
                        },
                        "description": "Event Id"
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "Start date in YYYY-MM-DD ",
                            "title": "Start Date"
                        },
                        "description": "Start date in YYYY-MM-DD "
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "End date in YYYY-MM-DD ",
                            "title": "End Date"
                        },
                        "description": "End date in YYYY-MM-DD "
                    },
                    {
                        "name": "slot",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/SlotEnum",
                            "description": "Slot of the facility (Morning/Afternoon/Night)"
                        },
                        "description": "Slot of the facility (Morning/Afternoon/Night)"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/events/facilities": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Check Facilities Available",
                "operationId": "check_facilities_available_events_facilities_get",
                "parameters": [
                    {
                        "name": "slot",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/SlotEnum",
                            "description": "Slot of the facility (Morning/Afternoon/Night)"
                        },
                        "description": "Slot of the facility (Morning/Afternoon/Night)"
                    },
                    {
                        "name": "venue_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "venue ID",
                            "title": "Venue Id"
                        },
                        "description": "venue ID"
                    },
                    {
                        "name": "band_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Band ID",
                            "title": "Band Id"
                        },
                        "description": "Band ID"
                    },
                    {
                        "name": "decoration_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Decoration ID",
                            "title": "Decoration Id"
                        },
                        "description": "Decoration ID"
                    },
                    {
                        "name": "no_days",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "description": "upto, how many days",
                            "default": 10,
                            "title": "No Days"
                        },
                        "description": "upto, how many days"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Body_update_banner_events_update_banner_patch": {
                "properties": {
                    "banner": {
                        "type": "string",
                        "format": "binary",
                        "title": "Banner",
                        "description": "Update event Banner image if exist it overwrites"
                    }
                },
                "type": "object",
                "required": [
                    "banner"
                ],
                "title": "Body_update_banner_events_update_banner_patch"
            },
            "EventBase": {
                "properties": {
                    "event_name": {
                        "type": "string",
                        "title": "Event Name"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "slot": {
                        "type": "string",
                        "enum": [
                            "Morning",
                            "Afternoon",
                            "Night"
                        ],
                        "title": "Slot"
                    },
                    "event_date": {
                        "type": "string",
                        "format": "date",
                        "title": "Event Date"
                    },
                    "venue_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Venue ID"
                    },
                    "band_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Band ID"
                    },
                    "decoration_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Decoration ID"
                    },
                    "snacks_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Snacks ID"
                    },
                    "snacks_count": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Snacks Count"
                    },
                    "ticket_enabled": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Ticket Enabled",
                        "default": false
                    },
                    "platinum_ticket_price": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Platinum Ticket Price"
                    },
                    "platinum_ticket_count": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Count of the platinum ticket"
                    },
                    "gold_ticket_price": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Gold Ticket Price"
                    },
                    "gold_ticket_count": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Count of the gold ticket"
                    },
                    "silver_ticket_price": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Silver Ticket Price"
                    },
                    "silver_ticket_count": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Count of the Silver ticket"
                    },
                    "payment_type": {
                        "type": "string",
                        "enum": [
                            "Full payment",
                            "Half payment"
                        ],
                        "title": "Payment Type"
                    },
                    "payment_mode": {
                        "type": "string",
                        "enum": [
                            "UPI",
                            "Credit Card",
                            "Debit Card"
                        ],
                        "title": "Payment Mode"
                    },
                    "payment_amount": {
                        "type": "integer",
                        "title": "total amount"
                    }
                },
                "type": "object",
                "required": [
                    "event_name",
                    "description",
                    "slot",
                    "event_date",
                    "payment_type",
                    "payment_mode",
                    "payment_amount"
                ],
                "title": "EventBase"
            },
            "EventResponse": {
                "properties": {
                    "event_name": {
                        "type": "string",
                        "title": "Event Name"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "slot": {
                        "type": "string",
                        "enum": [
                            "Morning",
                            "Afternoon",
                            "Night"
                        ],
                        "title": "Slot"
                    },
                    "event_date": {
                        "type": "string",
                        "format": "date",
                        "title": "Event Date"
                    },
                    "venue_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Venue ID"
                    },
                    "band_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Band ID"
                    },
                    "decoration_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Decoration ID"
                    },
                    "snacks_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Snacks ID"
                    },
                    "snacks_count": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Snacks Count"
                    },
                    "ticket_enabled": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Ticket Enabled",
                        "default": false
                    },
                    "platinum_ticket_price": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Platinum Ticket Price"
                    },
                    "platinum_ticket_count": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Count of the platinum ticket"
                    },
                    "gold_ticket_price": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Gold Ticket Price"
                    },
                    "gold_ticket_count": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Count of the gold ticket"
                    },
                    "silver_ticket_price": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Silver Ticket Price"
                    },
                    "silver_ticket_count": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Count of the Silver ticket"
                    },
                    "payment_type": {
                        "type": "string",
                        "enum": [
                            "Full payment",
                            "Half payment"
                        ],
                        "title": "Payment Type"
                    },
                    "payment_mode": {
                        "type": "string",
                        "enum": [
                            "UPI",
                            "Credit Card",
                            "Debit Card"
                        ],
                        "title": "Payment Mode"
                    },
                    "payment_amount": {
                        "type": "integer",
                        "title": "total amount"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "user_id": {
                        "type": "integer",
                        "title": "User Id"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date",
                        "title": "Created At"
                    },
                    "status": {
                        "$ref": "#/components/schemas/EventStatus"
                    },
                    "payment_id": {
                        "type": "integer",
                        "title": "Payment Id"
                    }
                },
                "type": "object",
                "required": [
                    "event_name",
                    "description",
                    "slot",
                    "event_date",
                    "payment_type",
                    "payment_mode",
                    "payment_amount",
                    "id",
                    "user_id",
                    "created_at",
                    "status",
                    "payment_id"
                ],
                "title": "EventResponse"
            },
            "EventStatus": {
                "type": "string",
                "enum": [
                    "booked",
                    "rescheduled",
                    "cancelled"
                ],
                "title": "EventStatus"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "SlotEnum": {
                "type": "string",
                "enum": [
                    "Morning",
                    "Afternoon",
                    "Night"
                ],
                "title": "SlotEnum"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}